Rename test files & classes
RE #37292
