+++ qt/scientific_interfaces/ISISReflectometry/test/ReflMockObjects.h
@@ -28,7 +28,7 @@
 #include "GUI/Runs/ISearchModel.h"
 #include "GUI/Runs/ISearcher.h"
 #include "GUI/Runs/SearchCriteria.h"
-#include "GUI/Save/IAsciiSaver.h"
+#include "GUI/Save/IFileSaver.h"
 #include "GUI/Save/ISavePresenter.h"
 #include "MantidAPI/AlgorithmManager.h"
 #include "MantidFrameworkTestHelpers/FallbackBoostOptionalIO.h"
@@ -329,12 +329,12 @@ public:
 };
 
 /**** Saver ****/
-class MockAsciiSaver : public IAsciiSaver {
+class MockFileSaver : public IFileSaver {
 public:
   MOCK_CONST_METHOD1(isValidSaveDirectory, bool(std::string const &));
   MOCK_CONST_METHOD4(save, void(std::string const &, std::vector<std::string> const &, std::vector<std::string> const &,
                                 FileFormatOptions const &));
-  virtual ~MockAsciiSaver() = default;
+  virtual ~MockFileSaver() = default;
 };
 
 /**** Job runner ****/
